Extract Transformer::DSL into a separate file
